shield is done
just as a side note the 3d team didn't think the shield needed to deplete stamina, but I strongly dissagree so while it does not take away the players stamina there is working variable in the ShieldController script where we can adjust the speed of depletion.

next I will start working the taser.