Fixed issue with money updating, removed the call to the Sell function from the ShopController script
Fixed an issue where the player's money would not update when returning to the game after purchasing something from the shop scene and leaving. I accomplished this by setting the variable "money" equal to "funds" before leaving the Shop scene, rather than when loading the Inventory scene. Additionally, I removed a call to the Sell function from the ShopController script.